+++ app/grid/GridTableFilter.cpp
@@ -1,0 +1,8 @@
+﻿#include "grids/GridTableFilter.h"
+
+
+namespace ctb::app
+{
+
+
+} // ctb::app
